Added in awe sensei and edited text
