--- Assets/DialogueController.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 09890c720ff694287b3163e93dc7c603
-timeCreated: 1494110406
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
