+++ Assets/Sounds/victorySong.wav.meta
@@ -1,0 +1,22 @@
+fileFormatVersion: 2
+guid: 55cc1bf9975ca46fc9ee51484dea81ae
+timeCreated: 1494277114
+licenseType: Free
+AudioImporter:
+  serializedVersion: 6
+  defaultSettings:
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  preloadAudioData: 1
+  loadInBackground: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
